formProcessing/requirements.txt
cairocffi==0.9.0
CairoSVG==2.1.3
cffi==1.11.5
click==6.7
cssselect2==0.2.1
defusedxml==0.5.0
Flask==1.0.2
html5lib==1.0.1
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
pdfrw==0.4
Pillow==5.2.0
pycparser==2.18
Pyphen==0.9.5
six==1.11.0
tinycss2==0.6.1
WeasyPrint==0.42.3
webencodings==0.5.1
Werkzeug==0.14.1
celery==4.2.1
redis==2.10.6
python-dateutil==2.7.3
web3==4.6.0
py-solc==3.1.0
